pyproject.toml
[project]
name = "power-traffic"
version = "0.1.0"
description = "Centralized traffic load orchestrator for Windows hosts"
requires-python = ">=3.11"
dependencies = [
  "PyYAML>=6.0.1"
]

[project.scripts]
power-traffic = "power_traffic.main:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

